+++ .github/workflows/ci_workflows.yml
@@ -32,10 +32,10 @@ jobs:
             python: 3.6
             toxenv: py36-test-oldestdeps   
             
-          - name: Python 3.7 with numpy 1.16 and full coverage
+          - name: Python 3.7 with numpy 1.21 and full coverage
             os: ubuntu-latest
             python: 3.7
-            toxenv: py37-test-alldeps-numpy116-cov
+            toxenv: py37-test-alldeps-numpy121-cov
 
           - name: Python 3.8 with all optional dependencies (Windows)
             os: windows-latest
