+++ astrocut/utils/utils.py
@@ -6,7 +6,6 @@ import warnings
 import numpy as np
 
 from datetime import date
-from itertools import product
 
 from astropy import wcs
 from astropy.io import fits
@@ -163,7 +162,7 @@ def _build_astrocut_primaryhdu(**keywords):
     primary_hdu = fits.PrimaryHDU()
     primary_hdu.header.extend([("ORIGIN", 'STScI/MAST', "institution responsible for creating this file"),
                                ("DATE", str(date.today()), "file creation date"),
-                                ('PROCVER', __version__, 'software version')])
+                               ('PROCVER', __version__, 'software version')])
     for kwd in keywords:
         primary_hdu.header[kwd] = keywords[kwd]
 
@@ -210,57 +209,3 @@ def get_fits(cutout_hdus, center_coord=None, output_path=None):
     return cutout_hdulist
 
 
-def path_to_footprints(path, size, img_wcs, max_pixels=10000):
-    """
-    Given a path that intersects with a wcs footprint, return
-    one or more rectangles that fully contain that intersection 
-    (plus padding given by 'size') with each rectangle no more than 
-    max_pixels in size.
-    
-    Parameters
-    ----------
-    path : `~astropy.coordinate.SkyCoord`
-        SkyCoord object list of coordinates that represent a continuous path.
-    size : array
-        Size of the rectangle centered on the path locations that must 
-        be included in the returned footprint(s). Formatted as [ny,nx]
-    img_wcs : `~astropy.wcs.WCS`
-        WCS object the path intersects with. Must include naxis information.
-    max_pixels : int
-        Optional, default 10000. The maximum area in pixels for individual
-        footprints.
-        
-    Returns
-    -------
-    response : list
-       List of footprints, each a dictionary of the form:
-       {'center_coord': `~astropy.coordinate.SkyCoord`, 'size': [ny,nx]}
-    """
-    
-    x, y = img_wcs.world_to_pixel(path)
-    
-    # Removing any coordinates outside of the img wcs
-    valid_locs = ((x >= 0) & (x < img_wcs.array_shape[0])) & ((y >= 0) & (y < img_wcs.array_shape[1]))
-    x = x[valid_locs]
-    y = y[valid_locs]
-    
-    bounds_list = _get_bounds(x, y, size)
-
-    combined_bounds = list()
-    cur_bounds = bounds_list[0]
-    for bounds in bounds_list[1:]:
-        new_bounds = _combine_bounds(cur_bounds, bounds)
-        
-        if _area(new_bounds) > max_pixels:
-            combined_bounds.append(cur_bounds)
-            cur_bounds = bounds
-        else:
-            cur_bounds = new_bounds
-            
-    combined_bounds.append(cur_bounds)
-    
-    footprints = []
-    for bounds in combined_bounds:
-        footprints.append(_get_args(bounds, img_wcs))
-        
-    return footprints
