+++ tox.ini
@@ -1,8 +1,8 @@
 [tox]
 envlist =
     py{36,37,38}-test{,-alldeps,-devdeps}{,-cov}
-    py{36,37,38}-test-numpy{117,118,119}
-    py{36,37,38}-test-astropy{30,40,lts}
+    py{36,37,38}-test-numpy{118,119,121}
+    py{36,37,38}-test-astropy{40,41,lts}
     build_docs
     linkcheck
     codestyle
@@ -34,22 +34,22 @@ description =
     devdeps: with the latest developer version of key dependencies
     oldestdeps: with the oldest supported version of key dependencies
     cov: and test coverage
-    numpy117: with numpy 1.17.*
     numpy118: with numpy 1.18.*
     numpy119: with numpy 1.19.*
-    astropy30: with astropy 3.0.*
+    numpy121: with numpy 1.21.*
     astropy40: with astropy 4.0.*
+    astropy41: with astropy 4.1.*
     astropylts: with the latest astropy LTS
 
 # The following provides some specific pinnings for key packages
 deps =
 
-    numpy116: numpy==1.16.*
-    numpy117: numpy==1.17.*
     numpy118: numpy==1.18.*
+    numpy119: numpy==1.19.*
+    numpy121: numpy==1.21.*
 
-    astropy30: astropy==3.0.*
     astropy40: astropy==4.0.*
+    astropy41: astropy==4.1.*
     astropylts: astropy==4.0.*
 
     devdeps: git+https://github.com/numpy/numpy.git#egg=numpy
